Updated the C++ files to include "using namespace std;" as well as a few other minor changes.
The full input now goes to the BattleshipAI class' handler instead of just the message. This allows it to determine for itself whether to send a message privately to someone or to the entire channel. Also, the decision on including the namespace declaration was made due to the fact the chance of a namespace conflict is extremely unlikely due to the face these programs will be largely logic and will not need to include many other code sources or libraries.